Zoom is now linear.

+++ gui/gui.c
@@ -56,7 +56,7 @@ void render_text(SDL_Renderer *renderer, const char *text, int parent_width, int
     SDL_DestroyTexture(text_texture);
 }
 
-void initialize_slider(struct slider *slider, int x, int y, int w, int h, uint8_t min_value, uint8_t max_value, uint8_t initial_value)
+void initialize_slider(struct slider *slider, int x, int y, int w, int h, float min_value, float max_value, float initial_value)
 {
     slider->track.x = x;
     slider->track.y = y;
@@ -126,7 +126,7 @@ void detect_release_from_slider(SDL_Event *event, struct slider *slider)
     slider->is_dragging = false;
 }
 
-void detect_drag_slider(SDL_Event *event, struct slider *slider, uint8_t *pixel_size, void (*render)(void))
+void detect_drag_slider(SDL_Event *event, struct slider *slider, float *pixel_size, void (*render)(void))
 {
     if (!slider->is_dragging)
         return;
